+++ kiplot/misc.py
@@ -20,6 +20,7 @@ NO_YAML_MODULE = 15
 NO_PCBNEW_MODULE = 16
 CORRUPTED_PCB = 17
 KICAD2STEP_ERR = 18
+WONT_OVERWRITE = 19
 
 CMD_EESCHEMA_DO = 'eeschema_do'
 URL_EESCHEMA_DO = 'https://github.com/INTI-CMNB/kicad-automation-scripts'
@@ -32,3 +33,4 @@ URL_KIBOM = 'https://github.com/INTI-CMNB/KiBoM'
 CMD_IBOM = 'generate_interactive_bom.py'
 URL_IBOM = 'https://github.com/INTI-CMNB/InteractiveHtmlBom'
 KICAD2STEP = 'kicad2step'
+EXAMPLE_CFG = 'example.kiplot.yaml'