+++ kiplot/out_pdf_pcb_print.py
@@ -19,6 +19,8 @@ class PDF_Pcb_Print(BaseOutput):  # noqa: F821
         This output is what you get from the 'File/Print' menu in pcbnew. """
     def __init__(self, name, type, description):
         super(PDF_Pcb_Print, self).__init__(name, type, description)
+        # We need layers, so we define it
+        self._layers = None
         # Options
         with document:
             self.output_name = ''